requirements.txt
aiofiles==23.2.1
aiohttp==3.9.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
bidict==0.22.1
certifi==2023.11.17
click==8.1.7
dnspython==2.4.2
fastapi==0.104.1
fastapi-socketio==0.0.10
frozenlist==1.4.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.1
idna==3.4
ifaddr==0.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
markdown2==2.4.10
MarkupSafe==2.1.3
multidict==6.0.4
nicegui==1.4.2
orjson==3.9.10
pscript==0.7.7
pydantic==2.5.1
pydantic_core==2.14.3
Pygments==2.17.1
pymongo==4.6.0
python-dotenv==1.0.0
python-engineio==4.8.0
python-multipart==0.0.6
python-socketio==5.10.0
PyYAML==6.0.1
simple-websocket==1.0.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.22.0
uvloop==0.19.0
vbuild==0.8.2
watchfiles==0.21.0
websockets==12.0
wsproto==1.2.0
yarl==1.9.2
